🚀 Feature: Validate upload song inputs
